+++ CHANGELOG.md
@@ -8,6 +8,7 @@ FEATURES:
 * **New Resource:** `r/tfe_scim_group_mapping`: Adds a resource to map a SCIM group to a team on Terraform Enterprise. By @skj-skj [#2090](https://github.com/hashicorp/terraform-provider-tfe/pull/2090)
 * **New Resource:** `r/tfe_provider_set` for managing shared provider configurations for internal testing. NOTE: This resource is subject to change and has limited support in HCP Terraform. It is not available to public. By @mogrogan [#2086](https://github.com/hashicorp/terraform-provider-tfe/pull/2086)
 * **New Data Source:** `d/tfe_provider_set` for retrieving shared provider configurations for internal testing. NOTE: This data source is subject to change and has limited support in HCP Terraform. It is not available to public. By @mogrogan [#2091](https://github.com/hashicorp/terraform-provider-tfe/pull/2091)
+* **New Resource:** `r/tfe_project_notification_configuration`: Adds a resource for managing project notification configurations, by @jillirami ([#1958](https://github.com/hashicorp/terraform-provider-tfe/pull/1958))
 
 ENHANCEMENTS:
 * `d/tfe_team`: Expose SCIM attributes (`scim_linked`, `scim_group_name`, `scim_sync_paused`, `scim_updated_at`) as computed read-only fields. These are only populated when SCIM is enabled on the TFE instance [#2088](https://github.com/hashicorp/terraform-provider-tfe/pull/2088)